% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_siteBufferToMultiLineIntersection.R
\name{future_siteBufferToMultiLineIntersection}
\alias{future_siteBufferToMultiLineIntersection}
\title{Take the weighted.mean of an sf MULTILINESTRING variable inside a buffer around sites}
\usage{
future_siteBufferToMultiLineIntersection(
  sites,
  name_sites,
  buffer = 1000,
  sfLines,
  name_sfLines_variable,
  name_sfLines_geometry = "geometry",
  nearest = FALSE,
  n_cores = NULL,
  max_size = 2000 * 1024^2
)
}
\arguments{
\item{sites}{sf data.frame with site locations and names}

\item{name_sites}{character string for the column name that contains the site names/IDs in \code{sites}}

\item{buffer}{numeric value for the buffer zone around sites in meters. Defaults to 1000 m.}

\item{sfLines}{sf data.frame with the MULTILINESTRING and the variable of interest}

\item{name_sfLines_variable}{character string for the column name that contains the values of the variable of interest in \code{sfLines}}

\item{name_sfLines_geometry}{character string for the column name that contains the geometry column in \code{sfLines}}

\item{nearest}{logical to indicate if the function assigns the nearest MULTILINESTRING to that site instead of the default weighted average. (e.g. used for coastal materials)}

\item{n_cores}{number of cores to use for parallel processing. If NULL, uses all available cores minus one.}

\item{max_size}{maximum size of global variables allowed for future processing (in bytes). Default is 2000 MB.}
}
\value{
sf data.frame
}
\description{
If there is no MULTILINESTRING within the buffer zone, this function will assign the nearest MULTILINESTRING to that site
}
\examples{

future_siteBufferToMultiLineIntersection(
   sites = CCVI_RealPropertySites  |>
   filter(`Coastal or Inland`=="Coastal"),
   name_sites = "RPIMSSiteCode",
   sfLines = CANCOAST_TIDALRANGE_V6,
   name_sfLines_variable = "TideRange",
   n_cores = 4
   )
}
