% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{geometricMean}
\alias{geometricMean}
\title{Calculate the geometric mean}
\usage{
geometricMean(x, warn = TRUE)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric scalar, the geometric mean of x
}
\description{
Calculate the geometric mean
}
\examples{
geometricMean(c(1,2,3,4,5))
}
