% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_comps.R
\name{join_comps}
\alias{join_comps}
\title{join_comps}
\usage{
join_comps(x)
}
\arguments{
\item{x}{named list of indicator data.frames. The name of the data.frame in the list will be used as a prefix for the Value and Score columns. All data.frames must have a HarbourCode,Value, and Score columns.}
}
\value{
A single data.frame with HarbourCode and the renamed Value and Score columns from each indicator data.frame, prefixed by the indicator name.
}
\description{
Join multiple indicator data.frames into a single data.frame by HarbourCode.
}
\examples{
df1 <- data.frame(HarbourCode = c("H1", "H2"), Value = c(10, 20), Score = c(1, 2))
df2 <- data.frame(HarbourCode = c("H1", "H3"), Value = c(30, 40), Score = c(3, 4))
join_comps(list(Indicator1 = df1, Indicator2 = df2))
}
