% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformSkewness.R
\name{transformSkewness}
\alias{transformSkewness}
\title{Transforms skewness to be within -0.5 and 0.5}
\usage{
transformSkewness(v)
}
\arguments{
\item{data}{a vector of data}
}
\value{
vector of transformed data
}
\description{
This function looks at the skewness of a variable. If it is not within
-0.5 and 0.5 (i.e. normal distribution) it will determine the lambda
required to transform the data to have a skewness equal to that of a normal
distribution.
}
\examples{
\dontrun{
data <- c(1,5,6,1000,60000,60001, 60002)
tS <- transformSkewness(data=data)
}

}
